+++ authorization_server/src/main/java/eu/europa/ec/eudi/signer/r3/authorization_server/model/oid4vp/VerifierClient.java
@@ -16,11 +16,11 @@
 
 package eu.europa.ec.eudi.signer.r3.authorization_server.model.oid4vp;
 
+import eu.europa.ec.eudi.signer.r3.authorization_server.config.OID4VPConfig;
 import eu.europa.ec.eudi.signer.r3.authorization_server.model.exception.OID4VPException;
 import eu.europa.ec.eudi.signer.r3.authorization_server.model.oid4vp.variables.VerifierCreatedVariables;
 import eu.europa.ec.eudi.signer.r3.authorization_server.model.oid4vp.variables.VerifierCreatedVariables.VerifierCreatedVariable;
 import eu.europa.ec.eudi.signer.r3.authorization_server.model.exception.OID4VPEnumError;
-import eu.europa.ec.eudi.signer.r3.authorization_server.config.VerifierConfig;
 import eu.europa.ec.eudi.signer.r3.common_tools.utils.WebUtils;
 
 import java.io.IOException;
@@ -54,11 +54,11 @@ public class VerifierClient {
     private final String transaction_id = "transaction_id";
 
     private static final Logger log = LoggerFactory.getLogger(VerifierClient.class);
-    private final VerifierConfig verifierProperties;
+    private final OID4VPConfig oid4VPConfig;
     private final VerifierCreatedVariables verifierVariables;
 
-    public VerifierClient(VerifierConfig verifierProperties) {
-        this.verifierProperties = verifierProperties;
+    public VerifierClient(OID4VPConfig oid4VPConfig) {
+        this.oid4VPConfig = oid4VPConfig;
         this.verifierVariables = new VerifierCreatedVariables();
     }
 
@@ -103,10 +103,6 @@ public class VerifierClient {
         log.info("Encoded Request URI: "+encoded_request_uri);
         String client_id = responseFromVerifier.getString(this.client_id);
         log.info("Client Id: "+ client_id);
-        if(!client_id.contains(this.verifierProperties.getClientId())) {
-            log.error("Client Id Received different from Client Id expected");
-            throw new Exception(OID4VPEnumError.UNEXPECTED_ERROR.getFormattedMessage());
-        }
         String presentation_id = responseFromVerifier.getString(this.transaction_id);
         log.info("Transaction Id: "+presentation_id);
 
@@ -151,8 +147,6 @@ public class VerifierClient {
         log.info("Encoded Request URI: "+encoded_request_uri);
         String client_id = responseFromVerifier.getString(this.client_id);
         log.info("Client Id: "+ client_id);
-        if(!client_id.contains(this.verifierProperties.getClientId()))
-            throw new Exception(OID4VPEnumError.UNEXPECTED_ERROR.getFormattedMessage());
         String presentation_id = responseFromVerifier.getString(this.transaction_id);
         log.info("Transaction Id: "+presentation_id);
 
@@ -183,7 +177,7 @@ public class VerifierClient {
         // makes a request to the verifier
         HttpResponse response;
         try {
-            response = WebUtils.httpPostRequest(verifierProperties.getPresentationUrl(), headers, bodyMessage);
+            response = WebUtils.httpPostRequest(this.oid4VPConfig.getVerifier().getPresentationUrl(), headers, bodyMessage);
         } catch (Exception e) {
             log.error("An error occurred when trying to connect to the Verifier. {}", e.getMessage());
             throw new Exception("An error occurred when trying to connect to the Verifier");
@@ -296,7 +290,7 @@ public class VerifierClient {
     }
 
     private String getLinkToWallet(String request_uri, String client_id) {
-        return "eudi-openid4vp://" + verifierProperties.getAddress() + "?client_id=" +
+        return this.oid4VPConfig.getWallet().getScheme() + this.oid4VPConfig.getVerifier().getDomain() + "?client_id=" +
                 client_id + "&request_uri=" + request_uri;
     }
 
@@ -386,11 +380,11 @@ public class VerifierClient {
     }
 
     private String getUrlToRetrieveVPTokenWithResponseCode(String presentation_id, String nonce, String code) {
-        return verifierProperties.getPresentationUrl() + "/" + presentation_id + "?nonce=" + nonce + "&response_code=" + code;
+        return this.oid4VPConfig.getVerifier().getPresentationUrl() + "/" + presentation_id + "?nonce=" + nonce + "&response_code=" + code;
     }
 
     private String getUrlToRetrieveVPToken(String presentation_id, String nonce) {
-        return verifierProperties.getPresentationUrl() + "/" + presentation_id + "?nonce=" + nonce;
+        return this.oid4VPConfig.getVerifier().getPresentationUrl() + "/" + presentation_id + "?nonce=" + nonce;
     }
 
     public JSONObject validateDeviceResponse(String MSO_MDoc_Device_Response) throws OID4VPException{
@@ -401,7 +395,7 @@ public class VerifierClient {
 
         HttpResponse response;
         try{
-            response = WebUtils.httpPostRequest(this.verifierProperties.getValidationUrl(), headers, body);
+            response = WebUtils.httpPostRequest(this.oid4VPConfig.getVerifier().getValidationUrl(), headers, body);
         }
         catch (Exception e){
             log.error("An error occurred when trying to make a request to the Verifier. {}", e.getMessage());